+++ CHANGELOG.md
@@ -7,9 +7,10 @@
 
 Nearly everything has had changes. Porting will require some effort.
 
-- Quo now depends on literal, meaning attributes (options) to queries are typed and explicit
+- Quo now depends on `literal`, meaning attributes (options) to queries are typed and explicit
 - Composing query objects now allows you to compose query classes rather than just instances of query objects
-- MergedQuery is now ComposedQuery
+- `MergedQuery` is now `ComposedQuery`
+- `EagerQuery` & `LoadedQuery` are now simply `CollectionBackedQuery`
 - Drop support for Ruby <= 3.1 and Rails < 7.0
 - Gem is now a Rails engine and relies on autoloading